Rename kzg_prover to prover
