+++ README.md
@@ -4,7 +4,7 @@ A monorepo for Summa, zk proof of solvency protocol
 
 ### Subdirectories
 
-- `kzg_prover`: Halo2 based zk prover and KZG polynomial interpolation for Summa
+- `prover`: Halo2 based zk prover and KZG polynomial interpolation for Summa
 - `contracts`: Solidity smart contracts for Summa
 - `backend` : Rust API to interact with Summa
