+++ prover/bin/gen_commit_and_proofs.rs
