+++ prover/examples/chunked_univariate_grand_sum.rs
