+++ prover/prints/univariate-grand-sum-layout.png
