+++ prover/src/circuits/mod.rs
