+++ prover/src/circuits/univariate_grand_sum.rs
