+++ prover/src/circuits/utils.rs
