+++ prover/src/cryptocurrency.rs
