+++ prover/src/utils/operation_helpers.rs
